Add files via upload
R markdown and knitted markdown with examples of Factor Analysis. Initial example is from Johnson and Wichern 1992 Applied Multivariate Statistical Analysis. 3rd Ed. Prentice-Hall. Second example is a simulation in the context of genotypes being evaluated in different environments. Made this for my own use to try to better understand Factor Analysis and its relationship to PCA in this context.